requirements.txt
PySimpleGUI~=4.60.1
PyYAML~=6.0